requirements.txt
pandas
selenium
python-dotenv
requests
tqdm 
rocketreach
stripe
flask
